improve C99 designated initializers * nested structures * multidimensional arrays * close #925
