+++ cxx-squid/src/test/resources/parser/C/a.out
