+++ cxx-squid/src/test/resources/parser/own/misc.cc
@@ -39,19 +39,3 @@ int main(int argc, char** argv)
     
     return foo()->i;
 }
-
-//C-COMPATIBILITY: C99 designated initializers
-struct Point { int x, y; };
-Point p = { .y = 45, .x = 72 };
-int a[6] = { [4] = 29, [2] = 15 };
-int widths[200] = { [0 ... 9] = 1, [10 ... 99] = 2, [100] = 3 };
-
-// C-COMPATIBILITY: C99 compound literals
-void compoundLiteral()
-{
-  struct foo {int a; char b[2];} structure;
-  structure = (struct foo) {x + y, 'a', 0};
-
-  int * array;
-  array = (int []) { 1, 2, 3, 4, 5};
-}
